Adds change to test so comment_count is an INT

+++ models/model.js
@@ -10,7 +10,7 @@ exports.fetchCategories = () => {
 exports.fetchReviews = () => {
   let queryStr = `
     SELECT 
-    owner, title, reviews.review_id, category, review_img_url, reviews.created_at, designer, reviews.votes, COUNT(comments.review_id) AS comment_count FROM reviews
+    owner, title, reviews.review_id, category, review_img_url, reviews.created_at, designer, reviews.votes, COUNT(comments.review_id)::INT AS comment_count FROM reviews
     LEFT JOIN comments ON reviews.review_id = comments.review_id
     GROUP BY reviews.owner, reviews.title, reviews.review_id, category, review_img_url, reviews.created_at, designer, reviews.votes
     ORDER BY reviews.created_at DESC;
